test(startup-mode-composition): add BDD tests exercising all HTTP startup modes
- Introduce a comprehensive startup-mode composition BDD suite exercising all 16 HTTP ports.
- Add new flows module for flow execution helpers to support fixture and DB modes.
- Expand flow_support module to include validation baseline extraction and additional state snapshots.
- Extend tests/startup_mode_composition_bdd.rs with new scenario validations and assertions.
- Wire new route endpoints (user list, annotations, route submission) in test flows.
- Harden validation error envelope stability checking across fixture and DB startup modes.
- Remove stubbed/ignored DB-mode unit tests - covered fully by BDD suite.
- Adjust server config to allow dead code in test to support state builders composition unit tests.

This enables end-to-end validation of adapter selection determinism between fixture-fallback and embedded PostgreSQL-backed modes at HTTP boundary, ensuring behavioral consistency and schema-resilience.
