+++ backend/tests/user_state_profile_interests_startup_modes_bdd.rs
@@ -14,10 +14,6 @@ use support::profile_interests::{
 use support::{drop_table, handle_cluster_setup_failure};
 
 #[path = "../src/server/config.rs"]
-#[allow(
-    dead_code,
-    reason = "tests import ServerConfig from server_config for BDD startup-mode checks"
-)]
 mod server_config;
 pub(crate) use server_config::ServerConfig;
